Adjust NumericalityValidator
