+++ lib/seed_builder/validate.rb
@@ -1,4 +1,4 @@
 require "seed_builder/validate/base"
 require "seed_builder/validate/unique/string"
-require "seed_builder/validate/length"
-require "seed_builder/validate/numericality"
+require "seed_builder/validate/length_validator"
+require "seed_builder/validate/numericality_validator"
